Stub version of expandable search dialog
Introduced the SearchDialog component, which is just a plain rect for now.
The SearchDialog apparition on the screen is animated with a
translation-opacity-scale effect with origin in top right corner, where the
Search button is positioned.

Added icons/background.png and icons/serach.png resources, which are self made
and will be probably replaced.